backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
psycopg[binary]==3.2.9
python-dotenv==1.0.1
djangorestframework==3.15.2